Download, parse, and display recommendations
I have a GitHub repo called "recommendations" with a growing list of
things I recommend. It would be nice to put that on my website, but
I don't want to duplicate it. With an emphasis for blogging, it
shouldn't be a surprise that Gatsby has fantastic markdown support.

I've added some magic to automatically download the recommendations file
from GitHub, parse it as markdown, and display it as a static page on my
website.

It looks horrible at the moment, but I can fix that in later commits.