Replace helmet with native Gatsby head
Gatsby v5 supports the `Head` component in pages. This means I no longer
need react-helmet.

In order to make this work with the theme, I've moved the web font to
`fontsource` and used the manifest generator for the icon (which is
a kind of useless icon tbh, I should replace it).